Generalize checkout script
This generalizes and upgrades the checkout-omnibus-software.rb script to
checkout.rb. This allows us to easily checkout specific versions of upstream
gem repositories based on versions that are locked into Gemfile.lock. This
ensures we're using the correct version of the gem and allow us to git merge
all of our Cinc patches.

I added some simple arguments so that it's easier to work with.