Move away from using patch to using git merge
This should hopefully simplify the way we patch cinc by using the power of git
directly instead of creating patch files. That way we can just make MR's to our
fork of chef (or whatever repo) and then use git merge when we build.

The basic workflow is the following:
1. Clone a full copy of upstream chef repo
2. Add Cinc fork as a remote
3. Run git merge from stable/cinc branch

If this fails, we'll know quickly that we need to update something.

Other changes include:
- Create build.sh which will assist with testing and building. This only works
  on Linux hosts currently.
- Remove patches since we're using git now.