Refactor source tarballs
This resolves #26 by providing a chef-full-$version.tar.xz tarball which
includes both the cinc source and the omnibus-software source that was used to
build it. In theory, this should be all that's need to rebuild cinc.

This keeps the cinc-$version.tar.xz for the cinc source itself

Other changes:

- Moves the source files into it's own top level "source" directory to make it
  easier to find and use.
- Add tests for source package

Signed-off-by: Lance Albertson <lance@osuosl.org>